Change game startup hook for EO compatibility
Electricity Overhaul is adding a few complex transpilers,
which seems to "invalidate" other postfix/prefix patches.
Our hooks into GameStateManager were ignored, probably due
to the new transpiled IL code not having the proper hooks
added to actually call our postfix function.

This has a potential to fail in other cases, but that is
a problem for Electricity Overhaul. But I don't see any
way to avoid it or fix it effectively for every case.

The newly used hooks seem cleaner anyway :)